Add ByteBuddy in place of mockito to remove thread-dependency when mocking ProcessEnvironment
