+++ system-stubs-core/src/test/java/uk/org/webcompere/systemstubs/CatchSystemExitTest.java
@@ -244,24 +244,6 @@ class CatchSystemExitTest {
 			= new SecurityManagerMock();
 
 		@Test
-		void getInCheck_is_delegated_to_original_security_manager(
-		) throws Exception {
-			originalSecurityManager.inCheck = true;
-			AtomicBoolean inCheck = new AtomicBoolean();
-			SystemStubs.withSecurityManager(
-				originalSecurityManager,
-				() -> SystemStubs.catchSystemExit(
-					() -> {
-						inCheck.set(getSecurityManager().getInCheck());
-						//ensure that catchSystemExit does not fail
-						exit(ARBITRARY_STATUS);
-					}
-				)
-			);
-			assertThat(inCheck).isTrue();
-		}
-
-		@Test
 		void security_context_of_original_security_manager_is_provided(
 		) throws Exception {
 			Object context = new Object();
@@ -280,29 +262,6 @@ class CatchSystemExitTest {
 				)
 			);
 			assertThat(contextDuringExecution).hasValue(context);
-		}
-
-		@Test
-		void checkTopLevelWindow_is_delegated_to_original_security_manager(
-		) throws Exception {
-			originalSecurityManager.topLevelWindow = true;
-			Object window = new Object();
-			AtomicBoolean check = new AtomicBoolean();
-			SystemStubs.withSecurityManager(
-				originalSecurityManager,
-				() -> SystemStubs.catchSystemExit(
-					() -> {
-						check.set(
-							getSecurityManager().checkTopLevelWindow(window)
-						);
-						//ensure that catchSystemExit does not fail
-						exit(ARBITRARY_STATUS);
-					}
-				)
-			);
-			assertThat(check).isTrue();
-			assertThat(originalSecurityManager.windowOfCheckTopLevelWindowCall)
-				.isSameAs(window);
 		}
 
 		@Test
